Male and female winners treated seperately.
API to get the king and queen!

+++ KingsAndQueensHat/IPenalty.cs
@@ -2,7 +2,7 @@
 {
     public interface IPenalty
     {
-        int ScorePenalty(TeamSet teamSet);
+        double ScorePenalty(TeamSet teamSet);
 
         /// <summary>
         /// How important this penalty category is
